chore: use webpackChain.toString as common stringify method
