+++ packages/builder/builder-shared/src/config.ts
@@ -274,3 +274,15 @@ export async function getMetaTags(
 
   return generateMetaTags(metaOptions);
 }
+
+export async function stringifyConfig(config: unknown, verbose?: boolean) {
+  const { default: WebpackChain } = await import('../compiled/webpack-5-chain');
+
+  // webpackChain.toString can be used as a common stringify method
+  const stringify = WebpackChain.toString as (
+    config: unknown,
+    options: { verbose?: boolean },
+  ) => string;
+
+  return stringify(config as any, { verbose });
+}
