+++ packages/builder/builder-shared/tests/config.test.ts
@@ -1,5 +1,6 @@
-import { expect, it } from 'vitest';
-import { getExtensions } from '../src/config';
+import { expect, it, describe } from 'vitest';
+import webpack from 'webpack';
+import { getExtensions, stringifyConfig } from '../src/config';
 
 it('should get default extensions correctly', async () => {
   expect(getExtensions()).toEqual(['.js', '.jsx', '.mjs', '.json']);
@@ -49,4 +50,46 @@ it('should get extensions with prefix by target', async () => {
     '.node.json',
     '.json',
   ]);
+});
+
+describe('stringifyConfig', () => {
+  it('should stringify webpack config correctly', async () => {
+    const { DefinePlugin } = webpack;
+    const webpackConfig = {
+      mode: 'development',
+      plugins: [new DefinePlugin({ foo: 'bar' })],
+    };
+
+    expect(await stringifyConfig(webpackConfig)).toMatchSnapshot();
+  });
+
+  it('should stringify webpack config with verbose option correctly', async () => {
+    const { DefinePlugin } = webpack;
+    const webpackConfig = {
+      mode: 'development',
+      plugins: [
+        new DefinePlugin({
+          foo: 'bar',
+          baz() {
+            const a = 1;
+            return a;
+          },
+        }),
+      ],
+    };
+
+    expect(await stringifyConfig(webpackConfig, true)).toMatchSnapshot();
+  });
+
+  it('should stringify builder config correctly', async () => {
+    const builderConfig = {
+      tools: {
+        webpackChain(chain: any) {
+          chain.devtool('eval');
+        },
+      },
+    };
+
+    expect(await stringifyConfig(builderConfig)).toMatchSnapshot();
+  });
 });
